factor out reusable piece out of index.js in lib/
- so that karma-benchmark-plotly-reporter can reuse them
- update test mock accordingly
